Move helper and types to datasource
